requirements.txt
flask
flask-cors
pdfplumber
google-generativeai
gunicorn